chore: Organize profile data transfer objects
This commit moves profile-related DTOs into more specific subpackages and introduces `DivisionItem.kt`.

- Moved `ChangeUsernameRequest` and `ChangeUsernameResponse` to `com.mycheva.app.profile.main.data.dto.changeusername`
- Moved `ChangePasswordRequest` and `ChangePasswordResponse` to `com.mycheva.app.profile.main.data.dto.changepassword`
- Moved `LogoutResponse` to `com.mycheva.app.profile.main.data.dto`
- Created `DivisionItem.kt` in `com.mycheva.app.profile.main.data.dto.getuser`
