+++ app/src/main/java/com/mycheva/app/profile/main/data/dto/LogoutResponse.kt
@@ -1,4 +1,4 @@
-package com.mycheva.app.profile.main.data
+package com.mycheva.app.profile.main.data.dto
 
 import com.google.gson.annotations.SerializedName
