+++ app/src/main/java/com/mycheva/app/profile/main/data/dto/changepassword/ChangePasswordRequest.kt
@@ -1,4 +1,4 @@
-package com.mycheva.app.profile.main.data
+package com.mycheva.app.profile.main.data.dto.changepassword
 
 import com.google.gson.annotations.SerializedName
