+++ app/src/main/java/com/mycheva/app/profile/main/data/dto/getuser/DivisionItem.kt
@@ -1,0 +1,12 @@
+package com.mycheva.app.profile.main.data.dto.getuser
+
+import com.google.gson.annotations.SerializedName
+
+data class DivisionItem(
+
+    @field:SerializedName("name")
+    val name: String,
+
+    @field:SerializedName("id")
+    val id: Int
+)
